homework-1/bank_api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
JayDeBeApi==1.2.3
JPype1==1.6.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
slowapi==0.1.9
bcrypt==5.0.0
